+++ package.json
@@ -37,6 +37,7 @@
   "devDependencies": {
     "@types/chai": "^4.3.11",
     "@types/express": "^4.17.21",
+    "@types/lodash": "^4.17.16",
     "@types/mocha": "^10.0.9",
     "@types/multer": "^1.4.12",
     "@types/nock": "^11.1.0",
