+++ src/routes/index.ts
@@ -42,7 +42,7 @@ import {
   uploadFileBufferToMinio,
   uploadToMinio,
 } from '../utils/minioClient';
-const _ = require('lodash');
+import { formatYearAndDay, mergeObjectsByOuPe } from '../utils/files';
 
 // Constants
 const VALID_MIME_TYPES = ['text/csv', 'application/json'] as const;
@@ -462,19 +462,17 @@ routes.get('/predict-inverse', async (req, res) => {
 routes.get('/predict-from-csv', async (req, res) => {
   try {
     const chapUrl = process.env.CHAP_URL as string | undefined;
-    const locationsStr = req.query.locations as string | undefined;
+    const locations = req.query.locations as string[] | undefined;
 
     if (!chapUrl) {
       logger.error('Chap URL not set');
       return res.status(500).json(createErrorResponse('ENV_MISSING', 'Chap URL not set'));
     }
 
-    if (!locationsStr) {
+    if (!Array.isArray(locations) || locations.length === 0) {
       logger.error('Locations not set');
       return res.status(400).json(createErrorResponse('LOCATIONS_MISSING', 'Locations not set'));
     }
-
-    const locations = locationsStr?.split(',')
 
     const csvData = await fetchCsvData(locations);
 
@@ -497,12 +495,12 @@ routes.get('/predict-from-csv', async (req, res) => {
       const organization: ClickhouseOrganzation = {
         code: data.wid,
         name: data.woreda,
-        level: 2,
+        level: 2, // @todo: get correct level
         type: 'point',
-        coordinates: [[0, 0]],
+        coordinates: [[0, 0]], // @todo: find way to set empty coordinates
         longitude: +data.lon,
         latitude: +data.lat,
-        timestamp: new Date().getMilliseconds(),
+        timestamp: new Date().getMilliseconds(), // @todo: get correct timestamp format
       };
 
       return {
@@ -538,14 +536,14 @@ routes.get('/predict-from-csv', async (req, res) => {
       mergeObjectsByOuPe(populations),
     );
 
-    const modelPrediction = new ModelPredictionUsingChap(chapUrl, logger);
+    const modelPrediction = new ModelPredictionUsingChap(chapUrl as string, logger);
 
     // start the Chap prediction job
     const predictResponse = await modelPrediction.predict({ data: payload });
 
     if (predictResponse?.status === 'success') {
       // wait for the prediction job to finish
-      const predictionResults = (await new Promise((resolve, reject) => {
+      const predictionResults = (await new Promise((resolve, _reject) => {
         const interval = setInterval(async () => {
           const statusResponse = await modelPrediction.getStatus();
           if (statusResponse?.status === 'idle' && statusResponse?.ready) {
@@ -694,87 +692,5 @@ routes.post('/upload', upload.single('file'), async (req, res) => {
       );
   }
 });
-
-/**
- * Formats the year and day of the year into YYYYMM format.
- *
- * @param {number} year - The year (e.g., 2012).
- * @param {number} dayOfYear - The day of the year (e.g., 200).
- * @returns {string} - The formatted date in YYYYMM format (e.g., 201207). Returns NaN on error.
- */
-function formatYearAndDay(year: number, dayOfYear: number): string {
-  // Check for invalid input
-  if (typeof year !== 'number' || typeof dayOfYear !== 'number' ||
-      isNaN(year) || isNaN(dayOfYear) ||
-      year < 1 || dayOfYear < 1 || dayOfYear > 366) { //handles leap years
-    throw new Error('Invalid input: year and dayOfYear must be positive numbers.');
-  }
-
-  // Calculate the month.  We'll use a simplified approach, assuming
-  // 366 days max.  For more accurate calculations, especially with
-  // leap years, you'd need a more complex algorithm.
-  let month = Math.floor((dayOfYear - 1) / 30) + 1; // Simplified month calculation.
-
-    // Adjust for months with fewer than 31 days.
-    if (dayOfYear <= 31) {
-        month = 1;
-    } else if (dayOfYear <= 59) { //Up to end of Feb in non-leap year
-        month = 2;
-    }
-    else if (dayOfYear <= 90) { //Up to end of March
-        month = 3;
-    }
-    else if (dayOfYear <= 120) {
-        month = 4;
-    }
-    else if (dayOfYear <= 151) {
-        month = 5;
-    }
-    else if (dayOfYear <= 181) {
-        month = 6;
-    }
-    else if (dayOfYear <= 212) {
-        month = 7;
-    }
-     else if (dayOfYear <= 243) {
-        month = 8;
-    }
-    else if (dayOfYear <= 273) {
-        month = 9;
-    }
-    else if (dayOfYear <= 304) {
-        month = 10;
-    }
-    else if (dayOfYear <= 334) {
-        month = 11;
-    }
-    else {
-        month = 12;
-    }
-
-
-  // Ensure month is two digits
-  const monthString = month.toString().padStart(2, '0');
-
-  // Combine year and month.
-  const result = year.toString() + monthString;
-  return result;
-}
-
-function mergeObjectsByOuPe(data: Array<{organizational_unit: string; period: string; value: number}>) {
-  const res = _.chain(data)
-    .groupBy((item: any) => `${item.period}`) // Group by 'ou' and 'pe'
-    .map((group: any) => _.mergeWith({}, ...group, (objValue: any, srcValue: any) => {
-      // Customizer function to sum 'value' properties.
-      if (_.isNumber(objValue) && _.isNumber(srcValue)) {
-        return objValue + srcValue;
-      }
-      return srcValue; // Default behavior: overwrite other properties.
-    }))
-    .values()    // Convert the resulting object back to an array
-    .value();    // Resolve the lodash chain
-
-  return res;
-}
 
 export default routes;
